requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.9.1
    # via django
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
cachetools==6.1.0
    # via lib-rql
colorama==0.4.6
    # via pytest
coverage[toml]==7.10.2
    # via pytest-cov
dj-database-url==3.0.1
    # via -r requirements.in
django==5.2.4
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-jazzmin
    #   django-rql
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-jazzmin==3.0.1
    # via -r requirements.in
django-rql==4.4.1
    # via -r requirements.in
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
faker==37.5.3
    # via
    #   -r requirements.in
    #   faker-vehicle
faker-vehicle==0.2.0
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.1.0
    # via pytest
jsonschema==4.25.0
    # via drf-spectacular
jsonschema-specifications==2025.4.1
    # via jsonschema
lark-parser==0.11.0
    # via lib-rql
lib-rql==2.0.2
    # via django-rql
packaging==25.0
    # via
    #   gunicorn
    #   pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
psycopg2-binary==2.9.10
    # via -r requirements.in
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
pytest-cov==6.2.1
    # via -r requirements.in
pytest-django==4.11.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via lib-rql
python-decouple==3.8
    # via -r requirements.in
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
typing-extensions==4.14.1
    # via referencing
tzdata==2025.2
    # via
    #   django
    #   faker
uritemplate==4.2.0
    # via drf-spectacular
whitenoise==6.9.0
    # via -r requirements.in
